Reorganize Directory structure.
The app, env, and static directories now live side by side on the
guest machine within the /usr/local/otm folder. We previously had
problems with littering the /usr/local folder with multiple
directories that were part of this project.

Fixed vagrant shell provisioner to check if the share directory is
mounted before trying to unmount it.

Cleaned up and added comments to the fabric file.