Update templates to receive instance from the request.
Added additional context processor to expose the request to the templates.
In order to do this, we had to paste in the original context processors
that Django uses by default.

Changed views to stop passing in instance as a top-level object.

Updated templates to access request.instance instead of instance.