Provide a more complete hash for plots
By default the audit hash does not provide for foreign keys that
reference the given object. For instance if we had the following:

Tree(id, plot_id)
Plot(id)

Updating the tree's plot with "tree.plot = someoplot" would update the
hash for the tree but not the plot record.

Since we want the plot's hash to change whenever a tree is updated at
all we use the tree's hashes as well as the plots.