Don't call standard login hooks
Since users are "audited" they require a call to "save_with_user". The
default hook here attempts to do something like:

user.last_login = today()
user.save()

Which causes an exception

This fix isn't really permanent, but it does enable login with the
django admin. In the future, we'll have to figure out how to handle this.