Add user view and template
I have mounted the user page under both /users/{username} because users
are global to the application. I have also created a redirect view that
sends /{instanceid}/users/{username} to
/users/{username}/?instance_id={instanceid}.

We anticipate using url rewriting so that
  http://philly.opentreemap.com/
maps to
  http://opentreemap.com/1/
and therefore
  http://philly.opentreemap.com/users/jwalgran
would map to
  http://opentreemap.com/1/users/jwalgran
and then redirect to
  http://opentreemap.com/users/jwalgran/?instance_id=1

When viewing the user page "under" an instance, all the activity,
badges, etc. can be filtered by that instance. If there is no
instance_id query param, the activity would be unfiltered.