Add plot popup view and template
Generates the content for the popup shown on the map when the user
clicks a tree.

I named the view that routes to the tree detail page so that I could
easily refer to it in the plot popup template tag.

I did not spend a lot of time adding data-* attributes or style
to the markup because we are waiting for full HTML prototypes from
the designers.