Show popup when clicking plot on map
A little more complicated then I would like, but the UTFGrid
control only returns the interactivty properties for a feature
and you need both data and coordinates to show a popup. My
solution uses Bacon to zip a stream of popup content with a
stream of map click coordinates.

I refactored the existing UTF grid hover code to use the new
asEventStream method whic I added to OpenLayers.Control.UTFGrid.

The popup has a link to the plot detail page. I added the
plot ID to the plot detail page template to make the change more
visible in demos.