Merge prototype markup into templates
I pulled markup out of the files in /prototype and stitched it into
the existing templates.

While the templates are being rebuilt I created an instance_base.html
as a DRY place to build content that depends only on the instance.

The eco-benefit search returns markup for both the number of matching
trees and the actual benefits, but these two data sets are shown in
different places on the template. My solution uses jquery to pull the
markup apart and insert it into the appropriate DOM locations.

I did not bring over holder.js because it is a dev-only dependency and
adding libraries into our JS build process is non-trivial. I replaced it
with a static tree image.

In a future commit the login and registration forms need to be properly
subclassed to add additional styling and fields.