Allow each instance to set eco benefit currency values
Provides a model (BenefitCurrencyConversion) that stores the
eco-unit-to-currency-unit factors.

Note that these values should not actually change with localization. For
instance, if someone is the UK decides to use £ for their currency
unit, it should show as £ to everyone (or, show in a local currency but
apply the exchange rate)

If an instance doesn't specify a conversion set, currency amounts will
simply be hidden.