Set search values from query string
Update 'species' and 'location' text box/autocompletes with the data
specified in the query string.

In particular, a mechanism using a custom events is being added. Any
search widget should implement a custom event called `restore` that
accepts as extra data param the value to update to.

Updates when using `restore` should *not* trigger a new search