Add a field template tag
There are several places in the application where we need to create
in place edit forms and show and hide values and edit fields based on
user permissions. This template tag wraps up the logic of fetching
metadata about a field and passing into the context of a child template
to render the actual markup.