Make udfs use 'udf:' prefix explicitly
UDF names must now be prefixed explicitly in most places where they are used. An exception being when in UDF internals, like UDFModel().

This decision was mostly stylistic, but was prompted by a bug in the newly created audit._deserialize_value method. When getting the django field for an audit value, it was retrieving the HStoreField for the UDFModel, which would try to deserialize the entire hstore (k,v hash). By being explicit about udf fields, it's easier to control the field that they provide to the deserialization method.