Add support for multi-region eco benefits
Previously eco benefits were calculated based on a fixed region
parameter. However, we can use the plot's location to define exactly
which region the tree belongs to.

The data is pending, however so the initial migration creates an overlay
over the entire united states for the Piedmont region. When the data
become available we can replace this migration