Use byte strings so that django doesn't freak out
For some reason these tests just broke (which freaked me out). Debugging
didn't lead to the change the broke it but did lead to a root cause. In
particular, it appears that django was trying to de-serialize a request
body a byte stream.

Since we're trying to python 3 ready we've been using unicode
strings. Changing these to byte strings seems to fix the problem.