Fix pluralization. Closes #257.
There are two updates. The first, I think that the X out of Y should
always be plural:

```
Based on 0 out of 1 trees
```

The second, we were doing localization in the views, which made our
numbers into strings and tricked the template pluralizer.

Since templates can do their own localization, we moved it there- fixing
the issue.