Support for approving/rejecting existing audits
Currently we have pending audits, which represent changes that have not
yet been applied. Now we also have the concept of "reviewing" an audit
that has already been applied.

If the review is 'accepted' then a ref is attached to the audit with
"ReviewAccepted" and nothing else happens.

If the review is 'rejected' then a ref is attached to the audit with
"ReviewRejected" and the change is rolled back.