Point JSONField introspection rule to correct path
This bug was discovered when trying to create a migration on the new exporter django app for a model that had a ForeignKey reference to instance. The following error message was printed:

(env)vagrant@precise64:/usr/local/otm/app/opentreemap$ python manage.py schemamigrati
on exporter --initial
Creating __init__.py in '/usr/local/otm/app/opentreemap/exporter/migrations'...
 ! Cannot freeze field 'treemap.instance.config'
 ! (this field has class treemap.json_field.JSONField)

 ! South cannot introspect some fields; this is probably because they are custom
 ! fields. If they worked in 0.6 or below, this is because we have removed the
 ! models parser (it often broke things).
 ! To fix this, read http://south.aeracode.org/wiki/MyFieldsDontWork

It looks like this was introduced when the JSONField class was moved to its own file, breaking the filepath reference in the introspection rule.