Use single species table with instance and otm_code
The InstanceSpecies join table was never implemented. Abandoning the
initial design of using a join table to link instances to a master
species list:

  - Allows map owners to have full control over their species list
  - Keeps species query logic simpler
  - Allows for UDFs on instance species

As of this commit the species rows no longer have an itree code.
Storing an itree code with the species is not correct, because the
specific itree code used for a species depend on the climate region in
which the tree is located.

The "otm_code" column (renamed symbol column) is now a unique identifier
for a species the lookup dictionary in ecobenefits/__init__.py links
these otm codes to itree codes by region. This mapping was based on the
data from http://www.itreetools.org/streets/resources/Streets%20Species%20Codes.xls

The otm_code field was seeded with USDA code values, but we are not
guaranteeing they will always match. We are creating our own set of
species identfiers that we can change if and when we need to.