Add a list of species dictionaries to treemap
This is our master species list used to generate Species rows when
creating instances. This was derrived by mergeing all the regional
itree streets species lists then deduplicating and cleaning the result
so that each otm_code is associated with a unique combination of
species, genus, cultivar, and other values.

The source CSV is included along with the shell script used to transform
it into dictionary literals. It is easier to make edits to the species
list by editing the csv then regenerating the JSON.