Make the inline edit form return form and response
The plotMover relies on the inlineEditStream's saveOkStream
to return the PUT response, which contains the new geoRevHash that we
use to update tiles.

--- meanwhile ---

Many other parts of the app use saveOkStream to get the form data that
was updated (such as to update street view).

--- fight! ---

Since these two approaches incompatible, saveOkStream now returns an
object with *both* the PUT response and the form data

Fixes #576.