Update the display of audit events on plot detail
To allow more flexibility in styling I made the short_descr property of
and audit display only the action that was taken and not include the
user.

I used that flexibilty to add grouping of audits when building the
context for the plot detail page. Audits by the same user on the same
day are grouped together to avoid needlessly repeating the username and
date in the recent edits list.

I allowed for showing an an audit that would normally be displayed as
"set readonly to False" as the more humane "made the tree editable" by
adding an Auditable class method that gives models a chance to generate
a unique format string for specific fields.

Fixes #567
Fixes #568