DRY up references to the add tree URL hash
We funnel users into adding trees from many places in an effort to
maximize user contributions. As a result, construting the url to the add
tree mode of the map page was scattered in a few different files. I
unified this as best I could by moving the hash text out to settings.py
and then forwarding it through to the templates and the client-side code,
via the templated settings.js.

Fixes #635