Style the password reset process
To get the custom template to properly override the stock templates I
needed to make sure our app is installed before Django registration.
Adam changed the app load order in 69f709eb, but I talked to him and
said it was most likely ok (and more correct) to switch it back.

Fixes #710