Drop get_field_by_name to fix app in gunicorn
fixes a pre-deployment bug found where the app was working for unknown reasons during tests and with the django runserver but was breaking when run with gunicorn. The problem, in a nutshell, is that get_field_by_name does a full import process that chokes, for unknown reasons, on a downstream app. This whole problem is circumvented by not using get_field_by_name. see comment block for more details.