Sidebar button: 'Edit Details' -> 'Quick Edit'
fixes #782 on github.

Fixes ambiguity in the two 'Edit Details' buttons:
If you click on a plot/tree, an infowindow will pop up with an Edit Details button, and the sidebar accordion will also have an Edit Details button. One redirects to the plot_detail page, the other triggers the inlineEditForm. This is confusing behavior, seeing two buttons with the same text that do different things.