Drop "recent" from "recent edits"
The edits view is intended to be paged through, sorted and filtered so
including the word recent is a little misleading

I changed the name of the view function and the template name for
consistency.

I did not rename the recent_edits_report feature string since it
is an arbitrary name and would require a more complicated multi-repo
change.

I did not rename "recent edits" on the plot detail page because it
is truly a list of the 5 most recent edits.

Fixes #793