Really disable moving markers. Fixes #831.
Work around a leaflet issue with dragging handler. Normally we'd simply
use 'marker.dragging.enable/disable'. However, the dragging context gets
lost somewhere between "enableMoving" and "disableMoving"

Once lost we can't unbind the drag handler and we're stuck. The
workaround is to hang on to the most recently added context and disable
when needed.