Pass treeId through from plot_detail endpoint
Tree ID is always passed through (either as a number or null) to ajax callers of the plot_detail endpoint. plot.js listens for treeId values and puts them into the edit form.

This is required so that we can put the newly created tree id into the edit form. Having the id in the edit form, in turn, is required in order to be able to delete a tree after adding it, without reloading the page.