Handle collection UDF date fields
The UI was using a simple text box for date fields within collection UDF
fields. I added a date picker to the field and modified the backend to
accept a date without a time as a valid value.

NOTE: This is an intermediate improvement to the date handling on
collection UDF fields. It should be further improved by displaying dates
in a local-appropriate format but posting them to the server in the ISO
format the back end expects.