Show planting site instead of plot via filter
There are several places throughout the application when we need to
convert a model name into a title displayed on a page. We are trying
to replace all instances of "plot" with "planting site" so while I was
working on editing collection UDFs I created a reusable filter to
handle the conversion.