Add index for Audit.current_value. Fixes #981.
This index helps us optimize for the following query:

```sql
SELECT DISTINCT ON ("treemap_audit"."model_id") "treemap_audit"."model_id", "treemap_audit"."updated"
    FROM "treemap_audit"
    WHERE ("treemap_audit"."instance_id" = 12  AND
        "treemap_audit"."model" = Tree  AND
        "treemap_audit"."field" = plot  AND
        "treemap_audit"."current_value" = 143493 )
    ORDER BY "treemap_audit"."model_id" DESC, "treemap_audit"."updated" DESC
```

This query is used to get the tree history for a given plot. The main
sticking point was the whole table sequence scan. Adding the index
dropped the query time by about 20x.

With this new addition, the audit table index represents an overhead of
about 28% of the audit table size. Justin and I decided this is a
worthwhile price to pay for now.