Add "default_permission" field to Role model
When adding map features to an instance we want to create default permissions to edit their fields for all existing roles, but there was no way to tell what permissions to use. Now we can store a default permission on each role and use that.

Also run a data migration to set the default role on our 3 existing roles in all instances