Enable overloading and extending registration_form.html
Normally in Django you can only override a template in one app by putting
another template with the same folder structure in a different app which
comes before the first in the INSTALLED_APPS order.

By doing so however, you can no longer {% extend %} the original template,
which leads to a sad state of copying entire templates into multiple places
with small changes between them, something which is unfortunately encouraged
by the official Django docs.

The django-apptemplates plugin allows you to specify which app to load a
template from, regardless of it's location in INSTALLED_APPS.

This allows you to override a template, while still using {% extend %} so
it does not need to be copied wholesale.

In this specific instance, we are using it to override
registration/registration_form.html in a separate Django app, then
extending it using {% extend treemap:registration/registration_form.html %}