Send the whole query to the ecoservice
Before we were going through acrobatics to send the right query and
params over the wire. Since the ecoservice should never be public we can
just send it raw SQL (yea!) and remove a ton of support code that wasn't
needed.

In addition we are able to move region detection into python, which is
nice since it makes the Go codebase smaller