Include config and logo url in instance API response
The mobile apps need access to the custom color and logo settings for
the instance so I have added them to the instance info API endpoint. I
decided to only include white listed config values in the response to
avoid possibly leaking sensitive config data to anyone with an API
key. We do not have any sensitive data now, but the config hash is a
dynamic grab bag and it is safer to not worry about what could be in
it.