Remove empty values from clickedIdStream
fixes #1098 on github

clickedIdStream represents the id of other features clicked from the mapFeature page map, allowing you to jump to that other feature's page. However, all clicks to the map were producing values in the stream, even ones that were not tied to an actual feature.

This fixes #1098 on github because the clicking of another feature tries to jump to a different page, which is blocked/prompted by the statePrompter during edit mode.