Provide default resource basis for eco template
For some reason, template tags are not working correctly within the eco_benefits.html template. Non existent key values are not rendering with a 'default' filter. I thought that any failure would result in a comparitable falsy value, so you can say {{ dictionary.nonexistent_key|default:"empty" }}, but this appears not to work. It is unclear if {% with %} and {% blocktrans %} have anything to do with it.

To sidestep this issue, provide default values.