Change the API return schema for instance eco
Previously, eco benefits were returned with a plot as an array. Mobile
applications only needed to know the number of eco fields that would be
included. Now, the eco fields for a plot are returned as a dictionary,
so the instance API must return the keys needed to read from the
dictionary and the order in which those keys should be displayed. I
copied the key ordering from the current web app UI.