Pull time off datetime before lexico comp to date
fixes #1337 on github

For hstore queries, we compare isoformatted date strings lexicographically. This does not work when comparing isoformatted datetimes to dates. For example, If you set the minimum bound on a search to "2014-03-24 00:00:00", you'll never capture dates in the database stored as "2014-03-24" using this method.

This fix forces datetimes to be represented as date strings.