Remove old fields when transforming api request
If the old 'firstname' and 'lastname' fields exist in the request after
the transformation wrapper the request handling fails with "unknown
field" errors. The goal of the transformation is to make v2 API requests
indistinguishable from v3 API requests, so it makes sense to remove the
now invalid keys from the request dict.