Try alternate mount syntax in vagrant provisioning
Laurence's linux box, my OS X machine, and Christopher's OS X machine
all had problems with remounting the shared folder inside the vagrant
box as the 'otm' user. Googling the issue turned up an alternate syntax
that works for us, but fails on Steve's linux box. I assume this has
something to due with versioning of either Vagrant or Virtual Box.

This change to the provisioning script tries the original mount method and
falls back to the second on failure.