Add warning to api.
Since this endpoint doesn't currently returned structured data about the failure
we need to decide what message to show the user based on the body text of the
message. This means that changing this string in the api could cause failures.
Until this is fixed this warning will hopefully prevent changes that might have
such an effect.