Pass region code, not region object (and rename variable)
Not sure how this ever worked, or why changing to Django 1.6 would have affected it.

Lower down this gets converted to a string using `str(...)`, and apparently that call used to extract the `ITreeRegion.code` field. But I see no method on `ITreeRegion` that would have accomplished that.

Fixes #1472

rename