Downgrade South back to 0.7.6
I originally upgraded South to version 0.8.4 because a weird error was
surfacing while I was writing new migrations. I was getting crashes with
the message "AttributeError: 'DateTimeField' object has no attribute
'model'" while adding `DateField` fields to a model. I found a
stackoverflow post (linked below) that mentions this error. There was no
accepted fix other than to upgrade South to 0.8.4. So I did, and this
error went away.

In another context that I believed to be unrelated, the upgrade to South
0.8.4 began causing a crash in a `treemap` migration on a freshly
formatted database: "django.db.utils.ProgrammingError: column
"mapfeature_ptr_id" is in a primary key". This error also turned up on
our continuous integration server once changes were merged into
master. It was eventually discovered that this was caused by the upgrade
of South.

By the grace of the spirit inside the computer, the original error that
prompted this upgrade is no longer replicable. It must have disappeared
while refactoring/squashing migrations.

So, to fix the second, crucial error, downgrade south back to 0.7.6.