Return instance bbox radius from API.
The iOS application uses the size of the instance to filter out search results
but this value was being specified at by the iOS app and was a hardcoded value
that was not in any way related to the actual instance. This was found to be a
problem because the radius accross LA is much larger than the radius that the
iOS app had hardcoded. Since this may be a useful feature in other applications
or contexts and since we do most geoprocessing at the API level anyway, we
calculate the radius on the server and provide it with the other data about the
instance.