Hide audits created by the system user.
The OTM1 -> OTM2 migrater system saves models using the system user.
This is because the OTM1 audit system is not robust enough to fully
migrate over audits, and we would like every field to have at least one
audit.

Because of this, we don't want to show system user audits in the UI, as it
would clutter and hide the "real" edits.