Fix audit records that crash the summary page
This commit does a lot:
Since species has a unicode method, we can delete the template code that
reads the scientific name and just print the species as is. This enables
us to also support cases where the value we pass is not a species object
with a valid `.scientific_name`, but a simple unicode object, as is the
case for migrated otm1 models.