Fix circular dependency problem with django apps
It looks like translation is importing *all other* django apps in the project
when it is used from treemap. This means that it will load apps that
depend on treemap when it is not finished import treemap. So while it
appears that treemap/otm1_migrator/otm_comments have sane, non-circular
dependencies on each other, the translation app is causing the circle.

I'm pretty sure this is actually a django pattern. Django enjoys
including dynamic apps that walk through installed apps and do magic
stuff. To compensate, they provide this alternative, string-based import
strategy that dynamic apps adhere to.