Prevent pre-populating date fields
On the tree detail page, the actions edit->save->edit would cause any
date fields on the page to be populated with the current date. This was
caused by the fact that calling "getDate" on the date picker will always
return a value, even if the text box to which the date picker is
connected is empty.

Fixes #1581