`getStateFromCurrentUrl` must include all possible keys
When initializing state from a URL include keys for all possible fields,
even those not present in the query. Modules initializing themselves from state
change stream need to be triggered even for empty values.

When this ability was recently lost (in an excellent refactor), the initial state lacked a
"mode" key which the treemap page depended on to set the initial model correctly.
Trees were not selectable, and UI tests failed.