Fix testem file watching
We were using too general of a globbing pattern, so it included the test
bundle as part of the watched files.

This caused testem to go into an infinite loop after the first file change.

By being more restrictive with the glob pattern, we ignore the /static/
directory, and thus avoid the infinite loop.