Parameterize 400 validation catchall
A view that uses this decorator was attempting to structure its response
in the format:
{'validationErrors': {<name_of_field>: [...]}}

But the view doesn't talk to the decorator, it only has the power to
raise a ValidationError and hope it will be swallowed
properly. Unfortunately ValidationError does not support this type of
nesting, so it must be done as a last step before returning the
response.

This can be seen as a stopgap alternative to refactoring the way all
endpoints and consuming clients wrap and unwrap validation errors. This
would require changing the app as well as the mobile clients, which is
not feasible. This seems like the best solution, all things considered.