Stopgap change in how "choice" fields are processed
The importer's code for validation of choice fields used `settings.CHOICES`,
an OTM1 variable not present in OTM2. So we need a new strategy.

For trees the choice fields will be UDF's so we'll use their choices.

For species we need a solution for "Flowering Period" and "Fruit or Nut Period".
For now I've hard-coded the USA's 4 seasons, but we need a solution that supports i18n.