Convert species "choice" fields to unstructured text
The species fields "flowering period" and "fruit and nut period" were a choice field in OTM1,
with choices hardcoded to spring/summer/fall/winter. There are two problems with this:
* Hardcoded strings won't work for i18n
* The LA species import file has values like "spring,summer,fall"

The right solution is to support UDF's on Species objects, make these multi-select choice fields, and
let international users edit the values.

That's not going to happen soon. Switching to unstructured text seems like a reasonable solution for now,
since we don't actually do anything with these fields.

Also:
* Remove "choice fields" logic from the importer. Choice fields for trees will be UDFs and validated there.