Make GenericImportEvent.errors.data always a []
Forcing this to be a list greatly simplifies template rendering of these
errors. We can say if errors.data then make a <ul> of each member of
data. For row-level validation this is not so simple, because each error
type appears to have its own schema, but for small number of cases of
file-level errors, it's easy to force this structure.