Comment out importer support for read only trees
The backend does not yet support read only tree enforcement so, to avoid
confusion, we are removing it from the importer as an available field.

I used the same "TODO: READONLY" comment everywhere I removed code so all
the blocks can be easily found and restored when the feature is implemented.